restructure command
